[JAX SC] Fix type annotations for dtypes and padding
* Annotate `_make_shape_dtype_struct` with `DTypeLike`.
* Annotate `pad_value` with `ArrayLike`.
* Annotate initializer return with `int` and remove obsolete suppressions.

PiperOrigin-RevId: 982515966

+++ jax_tpu_embedding/sparsecore/lib/nn/tests/test_utils.py
@@ -82,9 +82,7 @@ def row_col_id_initializer(
   return init
 
 
-def row_id_with_offset_initializer_value(
-    offset_value: int, row: int
-) -> jnp.float32:  # pyrefly: ignore[not-a-type]
+def row_id_with_offset_initializer_value(offset_value: int, row: int) -> int:
   """Returns the value for row_col_id_initializer."""
   return offset_value + row
